Creating a file to hold the options.
